Moved tests to proper folder, fixed missing file errors
