+++ frontend/index.js
@@ -1,0 +1,1 @@
+src/index.js
