+++ frontend/package.json
@@ -25,7 +25,7 @@
 	"scripts": {
 		"start": "react-scripts start",
 		"build": "react-scripts build",
-		"test": "mocha --require ignore-styles --require babel-polyfill --require babel-core -w test/helpers/browser.js \"test/tests.js\"",
+		"test": "mocha --require ignore-styles --require babel-polyfill --require babel-core -w src/assets/test/browser.js \"tests.js\"",
 		"eject": "react-scripts eject"
 	},
 	"devDependencies": {
