+++ frontend/src/assets/test/browser.js
