+++ frontend/src/index.js
@@ -3,7 +3,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './assets/css/index.css'
 import App from './js/App'
-import registerServiceWorker from './js/registerServiceWorker'
+import registerServiceWorker from './src/js/registerServiceWorker'
 import {BrowserRouter} from 'react-router-dom'
 /* eslint-enable no-unused-vars */
